moved apimanager into api folder
